+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. From versio
 
 ### Changes
 
+- Add schema cache query logging on `log-query=true` and `log-level=info|debug` by @steve-chavez in #5285
 - Reduce JWT cache memory usage by @mkleczek in #5229
 
 ## [16.4] - 2026-09-24